Setup controller test double and update dummy app
The controller double is just the bare minimum needed to mix in the
different controller modules.

The dummy app now has two proper actions, "welcome" for unauthorized
requests and "secure_page" for authenticated requests.
