+++ spec/dummy/app/controllers/welcome_controller.rb
@@ -1,5 +1,9 @@
 class WelcomeController < ApplicationController
-  before_filter :authenticate_user!
+  skip_before_filter :authenticate_user!, only: [:index]
+
   def index
+  end
+
+  def secure_page
   end
 end
