+++ spec/dummy/config/routes.rb
@@ -1,4 +1,5 @@
 Rails.application.routes.draw do
   mount Kracken::Engine => "/auth"
   get "welcome/index"
+  get "welcome/secure_page"
 end
